+++ src/modules/coordToolkit/components/CoordToolkit.vue
@@ -375,7 +375,18 @@ export default {
                 const el = this.$refs[id];
 
                 if (el) {
-                    values.push(el.value);
+                    // Support InputText component or native input
+                    let val = "";
+
+                    if (el.$el && el.$el.querySelector) {
+                        const input = el.$el.querySelector("input");
+
+                        val = input ? input.value : "";
+                    }
+                    else if (el.value !== undefined) {
+                        val = el.value;
+                    }
+                    values.push(val);
                 }
             });
             if (this.currentProjection.projName === "longlat") {
@@ -473,12 +484,12 @@ export default {
                 <InputText
                     :id="'coordinatesEastingField'"
                     ref="coordinatesEastingField"
-                    :value="coordinatesEasting.value"
+                    v-model="coordinatesEasting.value"
                     :label="$t(getLabel('eastingLabel'))"
                     :placeholder="isEnabled('search') ? $t('common:modules.coordToolkit.exampleAcronym') + coordinatesEastingExample : ''"
-                    :input="(value) => onInputEvent(value, coordinatesEasting)"
                     :readonly="isEnabled('supply')"
                     :class-obj="{ inputError: getEastingError }"
+                    @input="(value) => onInputEvent(value, coordinatesEasting)"
                 >
                     <div
                         v-if="isEnabled('supply') && !isMobile && showCopyButtons"
@@ -522,12 +533,12 @@ export default {
                 <InputText
                     :id="'coordinatesNorthingField'"
                     ref="coordinatesNorthingField"
+                    v-model="coordinatesNorthing.value"
                     :label="$t(getLabel('northingLabel'))"
                     :placeholder="isEnabled('search') ? $t('common:modules.coordToolkit.exampleAcronym') + coordinatesNorthingExample : ''"
-                    :value="coordinatesNorthing.value"
-                    :input="(value) => onInputEvent(value, coordinatesNorthing)"
                     :readonly="isEnabled('supply')"
                     :class-obj="{ inputError: getNorthingError }"
+                    @input="(value) => onInputEvent(value, coordinatesNorthing)"
                 >
                     <div
                         v-if="isEnabled('supply') && !isMobile && showCopyButtons"
@@ -594,8 +605,8 @@ export default {
                 <InputText
                     :id="'coordinatesHeightLabel'"
                     :label="$t('common:modules.coordToolkit.heightLabel')"
-                    :value="$t(height)"
-                    :readonly="true"
+                    :model-value="$t(height)"
+                    readonly
                     :placeholder="$t('common:modules.coordToolkit.heightLabel')"
                 />
             </div>