+++ src/modules/statisticDashboard/tests/unit/components/StatisticDashboardLegend.spec.js
@@ -231,7 +231,7 @@ describe("src/modules/statiscticDashboard/components/StatisticDashboardLegend.vu
     });
     describe("Custom classification", () => {
         it("should show correct step values according to store", async () => {
-            const wrapper = shallowMount(StatisticDashboardLegend, {
+            const wrapper = mount(StatisticDashboardLegend, {
                 global: {
                     plugins: [store]
                 }
@@ -242,9 +242,11 @@ describe("src/modules/statiscticDashboard/components/StatisticDashboardLegend.vu
             wrapper.vm.setClassificationMode("custom");
             await wrapper.vm.$nextTick();
 
-            expect(wrapper.getComponent("#value-range1").vm.value).to.equal("0");
-            expect(wrapper.getComponent("#value-range21").vm.value).to.equal("10");
-            expect(wrapper.getComponent("#value-range2").vm.value).to.equal("10");
+
+            expect(wrapper.find("input#value-range1").element.value).to.equal("0");
+            expect(wrapper.find("input#value-range21").element.value).to.equal("10");
+            expect(wrapper.find("input#value-range2").element.value).to.equal("10");
+
         });
         it("should change value in store correctly when changed", async () => {
             const wrapper = mount(StatisticDashboardLegend, {
@@ -264,7 +266,7 @@ describe("src/modules/statiscticDashboard/components/StatisticDashboardLegend.vu
             expect(wrapper.vm.stepValues).to.deep.equal([3, 10]);
         });
         it("should show correct color value according to store", async () => {
-            const wrapper = shallowMount(StatisticDashboardLegend, {
+            const wrapper = mount(StatisticDashboardLegend, {
                 global: {
                     plugins: [store]
                 }
@@ -276,8 +278,8 @@ describe("src/modules/statiscticDashboard/components/StatisticDashboardLegend.vu
             wrapper.vm.setColorPalette([[255, 0, 0], [0, 255, 0]]);
             await wrapper.vm.$nextTick();
 
-            expect(wrapper.getComponent("#color-range1").vm.value).to.equal("#ff0000");
-            expect(wrapper.getComponent("#color-range2").vm.value).to.equal("#00ff00");
+            expect(wrapper.find("input#color-range1").element.value).to.equal("#ff0000");
+            expect(wrapper.find("input#color-range2").element.value).to.equal("#00ff00");
         });
         it("should change value in store correctly when changed", async () => {
             const wrapper = mount(StatisticDashboardLegend, {