+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 ### Changed
 - Orientation: Nearby: If in an iFrame and geolocation is not enabled through config parameter, geolocation option will not show.
 - VectorTile, WebGL: Feature detection for MouseHover/GFI uses forEachFeatureAtPixel to support WebGL-rendered layers and avoid issues with getFeaturesInExtent.
+- InputText: Replace input text with InputText.
+- HTML input elements: Replaced HTML input elements of type text with shared modules component InputText.
 
 ### Deprecated